\name{ecocropK-package}
\alias{ecocropK-package}
\docType{package}
\title{
  ECOCROP-K
}
\description{
  R package for ECOCROP-K
}
\details{
  R package for ECOCROP-K (2020-02-24)
  Model of Integrated Impact and Vulnerability Evaluation of climate Change (MOTIVE) project
}
\author{
  Kwang Soo Kim, Shinwoo Hyun
}
\references{
  Kim, H., Hyun, S. W., Hoogenboom, G., Porter, C. H., & Kim, K. S. (2018). Fuzzy Union to Assess Climate Suitability of Annual Ryegrass (Lolium multiflorum), Alfalfa (Medicago sativa) and Sorghum (Sorghum bicolor). Scientific reports, 8(1), 1-15.
}
\keyword{ package }
\examples{
  \dontrun{
     ?ecocropK
  }
}
