\name{fzycrop-package}
\alias{fzycrop-package}
\docType{package}
\title{
  FZYCROP
}
\description{
  R package for FZYCROP
}
\details{
  R package for FZYCROP (2020-02-24)

  Model of Integrated Impact and Vulnerability Evaluation of climate Change (MOTIVE) project
}
\author{
  Kwang Soo Kim, Shinwoo Hyun
}
\references{
  Kim, H., Hyun, S. W., Hoogenboom, G., Porter, C. H., & Kim, K. S. (2018). Fuzzy Union to Assess Climate Suitability of Annual Ryegrass (Lolium multiflorum), Alfalfa (Medicago sativa) and Sorghum (Sorghum bicolor). Scientific reports, 8(1), 1-15.
}
\keyword{ package }
